Refactor test logic, remove unused code
